still a little unhappy with the handler but trying to make it generic
the use of the classes is nice because it allows for better custom flows but the way
i did it also introduces some issues with keeping references around. maybe if they were
registered as classes and i had a class factory that set the request method for the generic
modal handler instead? then the instance would be the full flow and could hold references
and ditch them when needed