remove webpage model
While some might argue the webpage model is a good idea, allowing for easier grouping of common bookmarks and reducing caches
it doesn't really make much sense and only gets the in the way for the small size of tb. this is part one of two to remove it, adding
and explicit tie between caches and a bookmark and putting the uri onto the bookmarks table. the rake task needs to be ran before
the server is started up so this might include a small maintenance window addition to prevent access to the app until it's ran